backend/requirements.txt
Flask==2.1.0
Flask-Cors==3.0.10
pymongo==3.12.0
urllib3==1.26.8
validators==0.20.0
requests==2.25.1
requests-oauthlib==1.3.0
PyJWT==2.4.0
pytest==6.2.5
textblob==0.17.1
redis==4.5.0
sentence-transformers==2.2.2
sendgrid==6.10.0
